Python:Refactor search connectors,consolidate Bing and Google search implementations, remove deprecated files, and update test imports accordingly.
