+++ python/semantic_kernel/connectors/search/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from semantic_kernel.connectors.search import bing, brave, google
+
+__all__ = ["bing", "brave", "google"]
