--- python/semantic_kernel/connectors/search/bing/__init__.py
@@ -1,6 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-from semantic_kernel.connectors.search.bing.bing_search import BingSearch
-from semantic_kernel.connectors.search.bing.bing_web_page import BingWebPage
-
-__all__ = ["BingSearch", "BingWebPage"]
