--- python/semantic_kernel/connectors/search/google/__init__.py
@@ -1,11 +1,0 @@
-# Copyright (c) Microsoft. All rights reserved.
-
-from semantic_kernel.connectors.search.google.google_search import GoogleSearch
-from semantic_kernel.connectors.search.google.google_search_response import GoogleSearchResponse
-from semantic_kernel.connectors.search.google.google_search_result import GoogleSearchResult
-
-__all__ = [
-    "GoogleSearch",
-    "GoogleSearchResponse",
-    "GoogleSearchResult",
-]
